Chapter 2: I made my own version of Rock, Paper, Scissors game.
Automate The Boring Stuff (ATBS) was telling me to make this program, by following the book / tutorial to write the source code which is already there, NOT to do it on my own.

However, I wanted to challenge myself and try to do it BEFORE seeing the source. Initially, I made it to print just what ATBS was saying (rpsGameOwnVersion.py),
but then I've wanted to take it further, so that I could really play this.

So, I added a computer bot or something like that, so that I can play with the computer. The computer is choosing randomly (rock, paper or scissors) on his own.

Initially, I've had an extra variable in which I've stored a string based on user's choice (e.g. 'ROCK' for 'r') and I also compared the strings such as 'r' and 'p' instead 'ROCK' and 'PAPER'. But I got rid of them, thinking that maybe it could make the code cleaner or easier to read.

I've used only the knowledge which I've learnt until this point (Chapter 2). I've used something which wasn't covered by the book until now though and that's the slash in print (e.g.  print('You don\'t want')   ). I knew that I had to use the slash to be able to print with apostrophe, either from the last time I've started to learn Python (last year) or from the days when I was doing Android custom ROM development.

rpsGame.py has the source from ATBS which I was told to type.

UPDATE:

After I've made my rpsGameOwnV2.py, I've checked the ATBS source code and it looks like this program wasn't supposed to print only the above output with that specific user input. Well, nevermind, since my rpsGameOwnV2.py covers that.