Chapter 5: I made my own TicTacToe game.
I was reading Automate The Boring Stuff and I was at Chapter 5 (well, still am as of typing this, after I kinda finished this program).

I got to the section "A Tic-Tac-Toe Board" and again, I didn't scroll beyond the sentence which told me to open a file editor to write the source code for this program.

I stopped there and I started to make this program.

After I finished this, I took a look at the Chapter to see what source code it was telling me to write. And I was surprised to see that it wasn't actually a complete tic-tac-toe game that I was supposed to write.

I wrote this using the knowlege I've learn since the first chapter. However, I believe I forgot some things and it could be done in a much simpler and cleaner way. I have no doubts about that.

The main program is ticTacToeOwn.py and that's what should be ran.